# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Healthcare Monitoring Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with AI health assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReply"
...
  /ai/emergency-assessment:
    post:
      operationId: aiEmergencyAssessment
      tags: [ai]
      summary: Assess emergency risk from vitals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiEmergencyAssessmentInput"
      responses:
        "200":
          description: Emergency assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiEmergencyAssessmentOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/doctor-summary:
    post:
      operationId: aiDoctorSummary
      tags: [ai]
      summary: Generate doctor-ready clinical summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiDoctorSummaryInput"
      responses:
        "200":
          description: Doctor summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDoctorSummaryOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/scheme-recommendations:
    post:
      operationId: aiSchemeRecommendations
      tags: [ai]
      summary: Recommend healthcare schemes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSchemeRecommendationInput"
      responses:
        "200":
          description: Scheme recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSchemeRecommendationOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/hospital-recommendations:
    post:
      operationId: aiHospitalRecommendations
      tags: [ai]
      summary: Recommend hospitals for patient needs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiHospitalRecommendationInput"
      responses:
        "200":
          description: Hospital recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiHospitalRecommendationOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AiChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        context:
          type: ["string", "null"]
          description: Optional vitals context
    AiReply:
      type: object
      required: [reply, suggestions]
      properties:
        reply:
          type: string
        suggestions:
          type: array
          items:
            type: string
    AiVitalMeasurement:
      type: object
      required: [heartRate, systolicBp, diastolicBp, spo2, glucose, temperature]
      properties:
        heartRate:
          type: number
        systolicBp:
          type: number
        diastolicBp:
          type: number
        spo2:
          type: number
        glucose:
          type: number
        temperature:
          type: number
        status:
          type: string
        recordedAt:
          type: string
...
    AiPatientProfile:
      type: object
      properties:
        name:
          type: string
        age:
          type: number
        gender:
          type: string
        conditions:
          type: array
          items:
            type: string
        location:
          type: string
    AiEmergencyAssessmentInput:
      type: object
      required: [currentVitals, profile, historicalTrends]
      properties:
        currentVitals:
          $ref: "#/components/schemas/AiVitalMeasurement"
        profile:
          $ref: "#/components/schemas/AiPatientProfile"
        historicalTrends:
          type: array
          items:
            $ref: "#/components/schemas/AiVitalMeasurement"
    AiEmergencySeverity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    AiEmergencyAssessmentOutput:
      type: object
      required: [severity, riskScore, explanation, recommendedAction, confidence]
      properties:
        severity:
          $ref: "#/components/schemas/AiEmergencySeverity"
        riskScore:
          type: integer
        explanation:
          type: string
        recommendedAction:
          type: string
        confidence:
          type: number
        fallback:
          type: boolean
    AiDoctorSummaryInput:
      type: object
      required: [patientHistory, alerts]
      properties:
        patientHistory:
          type: object
          required: [name, recentVitals]
          properties:
            name:
              type: string
            conditions:
              type: array
              items:
                type: string
            recentVitals:
              type: array
              items:
                $ref: "#/components/schemas/AiVitalMeasurement"
        alerts:
          type: array
          items:
            type: object
            required: [type, severity, message]
            properties:
              type:
                type: string
              severity:
                type: string
              message:
                type: string
              createdAt:
                type: string
        previousAssessments:
          type: array
          items:
            type: string
    AiDoctorSummaryOutput:
      type: object
      required: [clinicalSummary, criticalAlerts, importantObservations, confidence]
      properties:
        clinicalSummary:
          type: string
        criticalAlerts:
          type: array
          items:
            type: string
        importantObservations:
          type: array
          items:
            type: string
        confidence:
          type: number
        fallback:
          type: boolean
    AiSchemeRecommendationInput:
      type: object
      required: [age, gender, incomeCategory, healthConditions, location]
      properties:
        age:
          type: integer
        gender:
          type: string
        incomeCategory:
          type: string
        healthConditions:
          type: array
          items:
            type: string
        location:
          type: string
        availableSchemes:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
              eligibility:
                type: string
    AiSchemeRecommendationItem:
      type: object
      required: [scheme, eligibilityReason, benefits, score]
      properties:
        scheme:
          type: string
        eligibilityReason:
          type: string
        benefits:
          type: string
        score:
          type: integer
    AiSchemeRecommendationOutput:
      type: object
      required: [recommendations, confidence]
      properties:
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/AiSchemeRecommendationItem"
        confidence:
          type: number
        fallback:
          type: boolean
    AiHospitalRecommendationInput:
      type: object
      required: [condition, severity, location, requiredSpecialty]
      properties:
        condition:
          type: string
        severity:
          $ref: "#/components/schemas/AiEmergencySeverity"
        location:
          type: string
        requiredSpecialty:
          type: string
        availableHospitals:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              city:
                type: string
              specialties:
                type: array
                items:
                  type: string
              emergencyAvailable:
                type: boolean
              availableBeds:
                type: integer
    AiHospitalRecommendationItem:
      type: object
      required: [name, specialtyMatch, reason, score]
      properties:
        name:
          type: string
        specialtyMatch:
          type: string
        reason:
          type: string
        score:
          type: integer
    AiHospitalRecommendationOutput:
      type: object
      required: [hospitals, confidence]
      properties:
        hospitals:
          type: array
          items:
            $ref: "#/components/schemas/AiHospitalRecommendationItem"
        confidence:
          type: number
        fallback:
          type: boolean