# lib/api-zod/src/generated/types/aiDoctorSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthcare Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AiVitalMeasurement } from './aiVitalMeasurement';

export interface AiDoctorSummaryInput {
  patientHistory: {
    name: string;
    conditions?: string[];
    recentVitals: AiVitalMeasurement[];
  };
  alerts: Array<{
    type: string;
    severity: string;
    message: string;
    createdAt?: string;
  }>;
  previousAssessments?: string[];
}

export interface AiDoctorSummaryOutput {
  clinicalSummary: string;
  criticalAlerts: string[];
  importantObservations: string[];
  confidence: number;
  fallback?: boolean;
}