# lib/api-zod/src/generated/types/aiEmergencyAssessment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthcare Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AiVitalMeasurement } from './aiVitalMeasurement';

export interface AiPatientProfile {
  name?: string;
  age?: number;
  gender?: string;
  conditions?: string[];
  location?: string;
}

export type AiEmergencySeverity = typeof AiEmergencySeverity[keyof typeof AiEmergencySeverity];


export const AiEmergencySeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL',
} as const;

export interface AiEmergencyAssessmentInput {
  currentVitals: AiVitalMeasurement;
  profile: AiPatientProfile;
  historicalTrends: AiVitalMeasurement[];
}

export interface AiEmergencyAssessmentOutput {
  severity: AiEmergencySeverity;
  riskScore: number;
  explanation: string;
  recommendedAction: string;
  confidence: number;
  fallback?: boolean;
}