# lib/api-zod/src/generated/types/aiHospitalRecommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthcare Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AiEmergencySeverity } from './aiEmergencyAssessment';

export interface AiHospitalRecommendationInput {
  condition: string;
  severity: AiEmergencySeverity;
  location: string;
  requiredSpecialty: string;
  availableHospitals?: Array<{
    name?: string;
    city?: string;
    specialties?: string[];
    emergencyAvailable?: boolean;
    availableBeds?: number;
  }>;
}

export interface AiHospitalRecommendationItem {
  name: string;
  specialtyMatch: string;
  reason: string;
  score: number;
}

export interface AiHospitalRecommendationOutput {
  hospitals: AiHospitalRecommendationItem[];
  confidence: number;
  fallback?: boolean;
}