# lib/api-zod/src/generated/types/aiSchemeRecommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthcare Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AiSchemeRecommendationInput {
  age: number;
  gender: string;
  incomeCategory: string;
  healthConditions: string[];
  location: string;
  availableSchemes?: Array<{
    title?: string;
    description?: string;
    eligibility?: string;
  }>;
}

export interface AiSchemeRecommendationItem {
  scheme: string;
  eligibilityReason: string;
  benefits: string;
  score: number;
}

export interface AiSchemeRecommendationOutput {
  recommendations: AiSchemeRecommendationItem[];
  confidence: number;
  fallback?: boolean;
}