# lib/api-zod/src/generated/types/aiVitalMeasurement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Healthcare Monitoring Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AiVitalMeasurement {
  heartRate: number;
  systolicBp: number;
  diastolicBp: number;
  spo2: number;
  glucose: number;
  temperature: number;
  status?: string;
  recordedAt?: string;
}